moved shared methods between api and application controller to a mixin
