+++ app/controllers/api/api_controller.rb
@@ -1,5 +1,5 @@
 class Api::ApiController < ActionController::Base
-
+  include ApiApplicationMixin
 
   private
 
@@ -52,10 +52,6 @@ class Api::ApiController < ActionController::Base
     @owner = @device || find_by_id(params[:user_id])
   end
 
-  def model_find(type)
-    [User, Developer].find { |model| model.name == type.titleize}
-  end
-
   def check_privilege
     if @device
       unless @device.privilege_for(@dev) == "complete"
@@ -76,35 +72,4 @@ class Api::ApiController < ActionController::Base
     request.headers['X-User-Token'] == auth_token
   end
 
-  def method_missing(method_sym, *arguments, &block)
-    method_string = method_sym.to_s
-    if /(?<resource>[\w]+)_exists\?$/ =~ method_string
-      resource_exists?(resource, arguments[0])
-    elsif params[:id] && /(?<actor>[\w]+)_owns_(?<resource>[\w]+)\?$/ =~ method_string
-      actor_owns_resource? actor, resource, params[:id]
-    else
-      super
-    end
-  end
-
-  def resource_exists?(resource, arguments)
-    model = resource.titleize.constantize
-    render status: 404, json: { message: "#{model} does not exist" } unless arguments
-    arguments
-  end
-
-  def req_from_coposition_app?
-    @from_copo_app ||= request.headers["X-Secret-App-Key"] == Rails.application.secrets.mobile_app_key
-  end
-
-  def actor_owns_resource?(actor, resource, id)
-    model = resource.titleize.constantize
-    resource = model.find(id)
-    if (model == Checkin || model == Permission && actor == 'user')
-      owner = resource.device.user
-    else
-      owner = resource.send(actor)
-    end
-    owner == send("current_#{actor}")
-  end
 end
