+++ app/controllers/application_controller.rb
@@ -2,19 +2,7 @@ class ApplicationController < ActionController::Base
   # Prevent CSRF attacks by raising an exception.
   # For APIs, you may want to use :null_session instead.
   protect_from_forgery with: :exception
-
-  def method_missing(method_sym, *arguments, &block)
-    method_string = method_sym.to_s
-    if params[:id] && /(?<actor>[\w]+)_owns_(?<resource>[\w]+)\?$/ =~ method_string
-      actor_owns_resource? actor, resource, params[:id]
-    else
-      super
-    end
-  end
-
-  def req_from_coposition_app?
-    @from_copo_app ||= request.headers["X-Secret-App-Key"] == Rails.application.secrets.mobile_app_key
-  end
+  include ApiApplicationMixin
 
   def invalid_payload(msg, redirect_path)
     if req_from_coposition_app?
@@ -28,25 +16,5 @@ class ApplicationController < ActionController::Base
   def current_user?(user_id)
     current_user == User.find(user_id)
   end
-
-  def model_find(type)
-    [User, Developer].find { |model| model.name == type.titleize}
-  end
-
-  protected
-
-    def actor_owns_resource?(actor, resource, id)
-      # Called from method_missing
-      # Usage: user_owns_device?
-      # Checks whether resource belongs to actor
-      model = resource.titleize.constantize
-      resource = model.find(id)
-      if (model == Checkin || model == Permission && actor == 'user')
-        owner = resource.device.user
-      else
-        owner = resource.send(actor)
-      end
-      owner == send("current_#{actor}")
-    end
 
 end
